+++ src/scripts/json-schema-generator.ts
@@ -1,8 +1,8 @@
 /*****
  License
  --------------
- Copyright © 2017 Bill & Melinda Gates Foundation
- The Mojaloop files are made available by the Bill & Melinda Gates Foundation under the Apache License, Version 2.0 (the "License") and you may not use these files except in compliance with the License. You may obtain a copy of the License at
+ Copyright © 2020-2025 Mojaloop Foundation
+ The Mojaloop files are made available by the Mojaloop Foundation under the Apache License, Version 2.0 (the "License") and you may not use these files except in compliance with the License. You may obtain a copy of the License at
 
  http://www.apache.org/licenses/LICENSE-2.0
 
@@ -10,17 +10,18 @@
 
  Contributors
  --------------
- This is the official list (alphabetical ordering) of the Mojaloop project contributors for this file.
+ This is the official list of the Mojaloop project contributors for this file.
  Names of the original copyright holders (individuals or organizations)
  should be listed with a '*' in the first column. People who have
  contributed from an organization can be listed under the organization
  that actually holds the copyright for their contributions (see the
- Gates Foundation organization for an example). Those individuals should have
+ Mojaloop Foundation for an example). Those individuals should have
  their names indented and be marked with a '-'. Email address can be added
  optionally within square brackets <email>.
 
- * Gates Foundation
- - Name Surname <name.surname@gatesfoundation.com>
+ * Mojaloop Foundation
+ - Name Surname <name.surname@mojaloop.io>
+
  * Modusbox
  - Vijay Kumar Guthi <vijaya.guthi@modusbox.com>
 
@@ -68,7 +69,9 @@ $RefParser.dereference(openApiFile, (err, openApiSpecObj) => {
     indexFileContent += '\n'
     indexFileContent += '\n' + 'export namespace Schemas {'
     Object.keys(openApiSpec?.components?.schemas).forEach((schemaName) => {
-      indexFileContent += '\n' + `  export const ${schemaName} = JsonSchemas.${schemaName}`
+      // Replace - with _
+      const newSchemaName = schemaName.replaceAll('-', '_')
+      indexFileContent += '\n' + `  export const ${newSchemaName} = JsonSchemas["${schemaName}"]`
     })
     indexFileContent += '\n' + '}'
     indexFileContent += '\n'
@@ -87,7 +90,9 @@ $RefParser.dereference(openApiFile, (err, openApiSpecObj) => {
     typesFileContent += '\n'
     typesFileContent += '\n' + 'export namespace Types {'
     Object.keys(openApiSpec?.components?.schemas).forEach((schemaName) => {
-      typesFileContent += '\n' + `  export type ${schemaName} = components['schemas']['${schemaName}']`
+      // Replace - with _
+      const newSchemaName = schemaName.replaceAll('-', '_')
+      typesFileContent += '\n' + `  export type ${newSchemaName} = components['schemas']['${schemaName}']`
     })
     typesFileContent += '\n' + '}'
     typesFileContent += '\n'