+++ src/sdk-scheme-adapter/v2_1_0/backend/openapi.ts
@@ -776,7 +776,7 @@ export interface components {
             fulfilment?: components["schemas"]["IlpFulfilment"];
             /** @description Transaction ID from the DFSP backend, used to reconcile transactions between the Switch and DFSP backend systems. */
             homeTransactionId: string;
-            transferState?: components["schemas"]["transferState"];
+            transferState?: components["schemas"]["TransferStateFromBackend"];
         };
         /**
          * @description Below are the allowed values for the enumeration - RECEIVED DFSP has received the transfer. - RESERVED DFSP has reserved the transfer. - COMMITTED DFSP has successfully performed the transfer. - ABORTED DFSP has aborted the transfer due a rejection or failure to perform the transfer.
@@ -915,30 +915,30 @@ export interface components {
          */
         PartySubIdOrType: string;
         /**
-         * ExtensionKey
+         * ExtensionKey_v2_1_0
          * @description Extension key.
          */
-        ExtensionKey: string;
+        ExtensionKey_v2_1_0: string;
         /**
          * ExtensionValue
          * @description Extension value.
          */
         ExtensionValue: string;
         /**
-         * Extension
+         * Extension_v2_1_0
          * @description Data model for the complex type Extension.
          */
-        Extension: {
-            key: components["schemas"]["ExtensionKey"];
+        Extension_v2_1_0: {
+            key: components["schemas"]["ExtensionKey_v2_1_0"];
             value: components["schemas"]["ExtensionValue"];
         };
         /**
          * ExtensionList
          * @description Data model for the complex type ExtensionList. An optional list of extensions, specific to deployment.
          */
-        ExtensionList: {
+        ExtensionList_v2_1_0: {
             /** @description Number of Extension elements. */
-            extension: components["schemas"]["Extension"][];
+            extension: components["schemas"]["Extension_v2_1_0"][];
         };
         /**
          * PartyIdInfo
@@ -949,7 +949,7 @@ export interface components {
             partyIdentifier: components["schemas"]["PartyIdentifier"];
             partySubIdOrType?: components["schemas"]["PartySubIdOrType"];
             fspId?: components["schemas"]["FspId"];
-            extensionList?: components["schemas"]["ExtensionList"];
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"];
         };
         /**
          * PartyName
@@ -1058,7 +1058,7 @@ export interface components {
         ErrorInformation: {
             errorCode: components["schemas"]["ErrorCode"];
             errorDescription: components["schemas"]["ErrorDescription"];
-            extensionList?: components["schemas"]["ExtensionList"];
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"];
         };
         mojaloopError: {
             errorInformation?: components["schemas"]["ErrorInformation"];
@@ -1078,7 +1078,7 @@ export interface components {
             geoCode?: components["schemas"]["GeoCode"];
             ilpPacket?: components["schemas"]["IlpPacket"];
             condition?: components["schemas"]["IlpCondition"];
-            extensionList?: components["schemas"]["ExtensionList"];
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"];
             lastError?: components["schemas"]["quoteError"];
         };
         /**
@@ -1107,7 +1107,7 @@ export interface components {
         individualTransferResult: {
             transferId: components["schemas"]["CorrelationId"];
             fulfilment?: components["schemas"]["IlpFulfilment"];
-            extensionList?: components["schemas"]["ExtensionList"];
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"];
             transferState?: components["schemas"]["TransferState"];
             lastError?: components["schemas"]["transferError"];
         };
@@ -1126,8 +1126,8 @@ export interface components {
             quoteId?: components["schemas"]["CorrelationId"];
             quoteResponse?: components["schemas"]["individualQuoteResult"];
             fulfil?: components["schemas"]["individualTransferResult"];
-            quoteExtensions?: components["schemas"]["ExtensionList"];
-            transferExtensions?: components["schemas"]["ExtensionList"];
+            quoteExtensions?: components["schemas"]["ExtensionList_v2_1_0"];
+            transferExtensions?: components["schemas"]["ExtensionList_v2_1_0"];
             lastError?: components["schemas"]["transferError"];
         };
         /** BulkTransactionResponse */
@@ -1139,7 +1139,7 @@ export interface components {
             options?: components["schemas"]["bulkTransactionOptions"];
             /** @description List of individual transfer result in a bulk transfer response. */
             individualTransferResults: components["schemas"]["bulkTransactionIndividualTransferResult"][];
-            extensions?: components["schemas"]["ExtensionList"];
+            extensions?: components["schemas"]["ExtensionList_v2_1_0"];
         };
         /**
          * CurrencyConverter
@@ -1166,6 +1166,16 @@ export interface components {
          */
         AuthenticationType: "OTP" | "QRCODE" | "U2F";
         /**
+         * TransferStateFromBackend
+         * @description Below are the allowed values for the enumeration.
+         *     - RESERVED - Next ledger has reserved the transfer.
+         *     - COMMITTED - Next ledger has successfully performed the transfer.
+         *     Note: There is no ABORTED state, http error response with proper mojaloop error code should be used to abort / reject a transfer.
+         * @example COMMITTED
+         * @enum {string}
+         */
+        TransferStateFromBackend: "RESERVED" | "COMMITTED";
+        /**
          * TransfersIDPutResponse
          * @description The object sent in the PUT /transfers/{ID} callback.
          */
@@ -1173,7 +1183,7 @@ export interface components {
             fulfilment?: components["schemas"]["IlpFulfilment"];
             completedTimestamp?: components["schemas"]["DateTime"];
             transferState: components["schemas"]["TransferState"];
-            extensionList?: components["schemas"]["ExtensionList"];
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"];
         };
         /**
          * TransfersPostRequest
@@ -1187,7 +1197,7 @@ export interface components {
             ilpPacket: components["schemas"]["IlpPacket"];
             condition: components["schemas"]["IlpCondition"];
             expiration: components["schemas"]["DateTime"];
-            extensionList?: components["schemas"]["ExtensionList"];
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"];
         };
         /**
          * QuotesIDPutResponse
@@ -1202,7 +1212,7 @@ export interface components {
             geoCode?: components["schemas"]["GeoCode"];
             ilpPacket: components["schemas"]["IlpPacket"];
             condition: components["schemas"]["IlpCondition"];
-            extensionList?: components["schemas"]["ExtensionList"];
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"];
         };
         /**
          * TransactionScenario
@@ -1287,7 +1297,7 @@ export interface components {
             geoCode?: components["schemas"]["GeoCode"];
             note?: components["schemas"]["Note"];
             expiration?: components["schemas"]["DateTime"];
-            extensionList?: components["schemas"]["ExtensionList"];
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"];
         };
         /**
          * FxMoney
@@ -1322,7 +1332,7 @@ export interface components {
             expiration: components["schemas"]["DateTime"] & unknown;
             /** @description One or more charges which the FXP intends to levy as part of the currency conversion, or which the payee DFSP intends to add to the amount transferred. */
             charges?: components["schemas"]["FxCharge"][];
-            extensionList?: components["schemas"]["ExtensionList"] & unknown;
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"] & unknown;
         };
         /**
          * FxQuotesPostBackendRequest
@@ -1365,7 +1375,6 @@ export interface components {
         };
         fulfilment: components["schemas"]["IlpFulfilment"] & unknown;
         completedTimestamp: components["schemas"]["DateTime"] & unknown;
-        conversionState: components["schemas"]["TransferState"] & unknown;
         /**
          * FxTransfersPostBackendResponse
          * @description The object sent as a response for the POST /fxTransfers request.
@@ -1375,9 +1384,10 @@ export interface components {
             homeTransactionId?: string;
             fulfilment?: components["schemas"]["fulfilment"];
             completedTimestamp?: components["schemas"]["completedTimestamp"];
-            conversionState: components["schemas"]["conversionState"];
-            extensionList?: components["schemas"]["ExtensionList"];
+            conversionState: components["schemas"]["TransferStateFromBackend"];
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"];
         };
+        conversionState: components["schemas"]["TransferState"] & unknown;
         /**
          * FxTransfersPutBackendRequest
          * @description PUT /fxTransfers/{commitRequestId} object
@@ -1388,7 +1398,7 @@ export interface components {
             fulfilment?: components["schemas"]["fulfilment"];
             completedTimestamp?: components["schemas"]["completedTimestamp"];
             conversionState: components["schemas"]["conversionState"];
-            extensionList?: components["schemas"]["ExtensionList"];
+            extensionList?: components["schemas"]["ExtensionList_v2_1_0"];
         };
         /**
          * fxTransfersPatchBackendRequest